src/requirements.txt
requests>=2.25.0,<3.0.0
psutil>=5.8.0,<6.0.0
pysmb==1.2.9.1
pyyaml>=5.1.0,<7.0.0
# 指定稳定的依赖版本以避免兼容性问题
# pysmb 1.2.9.1是已知在Docker环境中稳定工作的版本